mailagent/requirements.txt
fastapi==0.65.0
uvicorn==0.13.4
pydantic
PyJWT==1.7.1
bcrypt
Authlib==0.15.5
starlette-authlib==0.1.8
httpx==0.20.0
requests
itsdangerous
psycopg2
SQLAlchemy
